code/AI/streamlit/tutorial/requirements.txt
streamlit==1.31.0
anthropic==0.7.0
python-dotenv==1.0.0
pandas==2.1.1
numpy==1.26.0
matplotlib==3.8.0
plotly==5.18.0 